% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele2locus.R
\name{allele2locus}
\alias{allele2locus}
\title{format microsat data from allelic to locus format}
\usage{
allele2locus(genotypes, sep_alleles = "")
}
\arguments{
\item{genotypes}{microsatellite data.frame in allelic format (two columns per locus).}

\item{sep_alleles}{how should the two alleles be concatenated? (defaults to "")}
}
\description{
format microsat data from allelic to locus format
}
\details{
takes allelic microsatellite data (two columns per locus, individuals in rows) and
transforms it into a one column per locus format, whereby the alleles are concetenated
with a given separator.
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com)
}

