% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_rich.R
\name{allele_rich}
\alias{allele_rich}
\title{Calculates allelic richness with rarefaction}
\usage{
allele_rich(genotypes, nboot = 1000, nsamp = NULL, nloc = NULL)
}
\arguments{
\item{genotypes}{microsatellite genotypes in two-column per locus format}

\item{nboot}{number of bootstraps for rarefaction}

\item{nsamp}{number of individuals to resample, usually is smaller than the smallest population to compare}

\item{nloc}{number of loci to resample, usually is smaller than the number of loci of the population with
the fewest loci}
}
\description{
Calculates allelic richness with rarefaction
}
\examples{
data(microsat_data)
out <- allele_rich(microsat_data, nboot = 5, nsamp = 5, nloc = 5)
out

}
