% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_ratio.R
\name{m_ratio}
\alias{m_ratio}
\title{Approximate m-ratio}
\usage{
m_ratio(gtypes_geno, rpt_size = 8:2)
}
\arguments{
\item{gtypes_geno}{microsatellite genotypes formatted as gtypes object with strataG}

\item{rpt_size}{The size of potential repetitive units in a microsatellite. Default is from 8 to 2.}
}
\description{
Calculate Garza-Williamson M ratio (bottleneck) statistic for microsatellite data.
The function is based on mRatio() from the strataG package. However, instead of requiring
the same repeat length for all alleles within a locus, it calculates the m-ratio by deviding
the allele-range by the repeat lengths that is most common. Is is therefore an approximation
but maximises the amount of information used for non-optimally scored microsatellite datasets.
}
