% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_genepop.R
\name{make_genepop}
\alias{make_genepop}
\title{format allelic microsats to genepop}
\usage{
make_genepop(x)
}
\arguments{
\item{x}{genotypes}
}
\description{
outputs a data.frame with genepop format
}
\details{
so far, the input is the standard format
for the bottleneck analysis, i.e. first column "id",
second column "pop", third column "cluster", all
following columns are loci (two columns per locus).
"id" and "cluster" will be deleted and "pop" will
be kept for the formatting.
}
\author{
Emily Humble (emily.lhumble@gmail.com)
         Martin Stoffel (martin.adam.stoffel@gmail.com)
}
