% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsimr_data.R
\docType{data}
\name{microsimr_data}
\alias{microsimr_data}
\title{simulated microsatellite data from microsimr}
\format{An object of class \code{data.frame} with 40 rows and 17 columns.}
\usage{
microsimr_data
}
\description{
simulated microsatellite data from microsimr
}
\keyword{datasets}

