% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssumstats.R
\name{mssumstats}
\alias{mssumstats}
\title{takes output from microsimr and outputs summary statistics}
\usage{
mssumstats(simd_data)
}
\description{
takes output from microsimr and outputs summary statistics
}
\examples{

}

