% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssumstats_old.R
\name{mssumstats_old}
\alias{mssumstats_old}
\title{Takes output from microsimr or an empirical microsatellite dataset and outputs summary statistics}
\usage{
mssumstats_old(simd_data, type = c("microsimr", "microsats"), by_pop = NULL,
  start_geno = NULL, data_type = c("simulated", "empirical"),
  rarefaction = FALSE, nsamp = NULL, nloc = NULL, nboot = 1000)
}
\arguments{
\item{simd_data}{output from microsimr}

\item{type}{"microsimr" for microsimr output, "microsats" for allelic microsatellite format}

\item{by_pop}{name of population variable. If specified, all summary statistics will be calculated within populations}

\item{start_geno}{integer, specifying the first column with genotypes. If now specified, all column are expected to be genotypes}

\item{data_type}{if "empirical", the mratio will be calculated differently, see ?m_ratio}

\item{rarefaction}{if TRUE, calculates mean and sd number of alleles as mean of n_samp individuals and n_loc loci over n_boot bootstraps}

\item{nsamp}{number of samples to subssample}

\item{nloc}{number of loci to subsample}

\item{nboot}{number of bootstraps}
}
\description{
Takes output from microsimr or an empirical microsatellite dataset and outputs summary statistics
}
\examples{
data(microsimr_data)
mssumstats(microsimr_data)

data(microsat_data)




}
