% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssumstats_old3.R
\name{mssumstats_old3}
\alias{mssumstats_old3}
\title{Takes output from microsimr or an empirical microsatellite dataset and outputs summary statistics}
\usage{
mssumstats_old3(data, datatype = c("microsats", "microsimr"), by_pop = NULL,
  start_geno = NULL, mratio = c("strict", "loose"), rarefaction = FALSE,
  nsamp = NULL, nloc = NULL, nboot = 1000)
}
\arguments{
\item{data}{microsatellites, whereby every locus is represented by two adjacent columns. Or microsimr output.}

\item{datatype}{defaults to "microsats" for allelic microsatellite format, "microsimr" for microsimr output}

\item{by_pop}{name of population variable. If specified, all summary statistics will be calculated within populations}

\item{start_geno}{integer, specifying the first column with genotypes. If now specified, all column are expected to be genotypes}

\item{mratio}{defaults to "strict". if "loose", the mratio will be calculated differently, see ?m_ratio}

\item{rarefaction}{if TRUE, calculates mean and sd number of alleles as mean of n_samp individuals and n_loc loci over n_boot bootstraps}

\item{nsamp}{number of samples to subssample}

\item{nloc}{number of loci to subsample}

\item{nboot}{number of bootstraps}
}
\description{
Takes output from microsimr or an empirical microsatellite dataset and outputs summary statistics
}
\examples{
data(microsimr_data)
mssumstats(microsimr_data, datatype = "microsimr")

data(microsat_data)
mssumstats(microsat_data, datatype = "microsats")

data(fur_seal)
mssumstats(fur_seal, datatype = "microsats", by_pop = "pop", start_geno = 4)

}
