% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_sheets.R
\name{read_excel_sheets}
\alias{read_excel_sheets}
\title{loads multiple data frames into list from xlsx and assigns the
tab names from the spreadsheet to the list elements}
\usage{
read_excel_sheets(data_path)
}
\arguments{
\item{data_path}{path to excel file with multiple data.frames}
}
\description{
loads multiple data frames into list from xlsx and assigns the
tab names from the spreadsheet to the list elements
}

