% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dflist_to_xls.R
\name{write_dflist_to_xls}
\alias{write_dflist_to_xls}
\title{takes list of named data.frames and writes to xls with respective names as sheet names}
\usage{
write_dflist_to_xls(dflist, file_name)
}
\arguments{
\item{dflist}{list of named data.frames}

\item{file_name}{file name for xls file}
}
\description{
takes list of named data.frames and writes to xls with respective names as sheet names
}
